Update the install link in the documents to lead to github over bitbucket
As this is now hosted publically on Github i have just updated the
install guides. I decided as this is only going to be cloned i do not
need to add this to packagist. I don't really want people installing it
into their project but rather cloning then using as a base.